+++ src/assets/sass/website/components/services/service-details.scss
@@ -1,0 +1,18 @@
+@import "../../color-palette.scss";
+@import "../../variables.scss";
+@import "../../mixins.scss";
+
+.work-details-sections{
+    width: 80%;
+    margin: auto;
+    color: $header-item;
+    @include is-extra-small-mobile {
+      width: 90%;
+    }
+    @include is-mobile {
+      width: 90%;
+    }
+    &__section{
+        margin: 60px 0 60px 0;
+    }
+}
