--- src/assets/sass/website/components/work/work-sections-title.scss
@@ -1,4 +1,0 @@
-@import "../../mixins.scss";
-@import "../../color-palette.scss";
-@import "../../shared/half-clipped-shape.scss";
-
