+++ src/assets/sass/website/shared/WorkServiceSections/work-details-section.scss
@@ -13,18 +13,6 @@
   background-color: $accent;
 }
 .work-details-sections{
-    width: 80%;
-    margin: auto;
-    color: $header-item;
-    @include is-extra-small-mobile {
-      width: 90%;
-    }
-    @include is-mobile {
-      width: 90%;
-    }
-    &__section{
-        margin: 60px 0 60px 0;
-    }
     &__container{
       background-color: $dark-background;
       padding: 0 110px  0 110px ;
