+++ src/website/shared/WorkServiceSections/DetailsNormalSection.vue
@@ -139,5 +139,5 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "../../../assets/sass/website/components/work/work-details-section.scss";
+@import "../../../assets/sass/website/shared/WorkServiceSections/work-details-section.scss";
 </style>
